Move the frontend code into the frontend or shared folders
