+++ package.json
@@ -97,11 +97,11 @@
     "compile-uplot": "terser --module --ecma 2018 --compress --mangle -o ./resources/uPlot.esm.min.js -- node_modules/uplot/dist/uPlot.esm.js",
     "compile": "tsc",
     "postcompile": "npm run prep-static && npm run download-font",
-    "prep-static": "cp dist/src/views/*.js ./resources/",
+    "prep-static": "cp dist/src/frontend/*.js ./resources/",
     "download-font": "node dist/src/download.js https://github.com/googlefonts/roboto/releases/download/v2.136/roboto-hinted.zip tmp/roboto-hinted.zip && unzip -o -d dist tmp/roboto-hinted.zip",
     
     "start": "node --enable-source-maps --experimental-json-modules ./dist/src/index.js",
-    "nodemon": "DEV=true nodemon --enable-source-maps --experimental-json-modules ./dist/src/index.js --watch ./dist/src --watch ./dist/package.json --watch ./src/views/ --ext js,json,html",
+    "nodemon": "DEV=true nodemon --enable-source-maps --experimental-json-modules ./dist/src/index.js --watch ./dist/src --watch ./dist/package.json --ext js,json,html",
 
     "format": "prettier --config .prettierrc '{src,tests}/**/*.ts' --write",
     "verify": "npm run lint",
