+++ src/index.ts
@@ -134,7 +134,7 @@ router.post(
 
 if (DEV) {
   router.get(`${siteConfig.staticUrl}/:filename*`, serveStaticResource);
-  router.get(`/src/views/:filename*`, serveViewJs);
+  router.get(`/src/frontend/:filename*`, serveViewJs);
   router.get(
     `${siteConfig.reportsUrl}/:change/figure-html/:filename`,
     serveReport
